Add "Get Projects" and "Find By Id" from Production Service (#251)
* Implement service to get projects from the real service and to find a given project by ID.

Includes the refactors and improvements that get this done

* Refactor amount and fee rate types from ulong to long across multiple models and view models

* Refactor amount conversion and update UI bindings for BTC display

* Remove unnecessary ObserveOn calls

* Remove comment

* Move namespace

+++ src/Angor/Avalonia/AngorApp/Sections/Browse/StageDesign.cs
@@ -1,11 +1,11 @@
-using AngorApp.Model;
+using Angor.UI.Model;
 
 namespace AngorApp.Sections.Browse;
 
 public class StageDesign : IStage
 {
     public DateOnly ReleaseDate { get; set; }
-    public uint Amount { get; set; }
+    public long Amount { get; set; }
     public int Index { get; set; }
     public double Weight { get; set; }
 }
